Horizontal
Please check code

+++ Main/helpers.ino
@@ -5,7 +5,7 @@
   This function prints a single into to serial followed by a linebreak
 */
 
-void printIntToSerial(int valueToPrint) {
+void printIntToSerial(float valueToPrint) {
 
   Serial.print(valueToPrint);
   Serial.print("\n");
